# specs/001-create-a-web/contracts/api-spec.yaml
openapi: 3.0.0
info:
  title: Sydney Techno Event Analytics API
  version: 1.0.0
  description: API for uploading CSV data and retrieving event analytics
servers:
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /uploads:
    post:
      summary: Upload CSV file for processing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - platform
              properties:
                file:
                  type: string
                  format: binary
                  maxLength: 104857600  # 100MB
                platform:
                  type: string
                  enum: [resident_advisor, humanitix, moshtix]
      responses:
        '201':
          description: Upload created and processing started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Upload'
        '400':
          description: Invalid file or platform
        '413':
          description: File too large

    get:
      summary: List all uploads
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of uploads
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploads:
                    type: array
                    items:
                      $ref: '#/components/schemas/Upload'
                  total:
                    type: integer
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Upload:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        fileSize:
          type: integer
        platform:
          type: string
          enum: [resident_advisor, humanitix, moshtix]
        status:
          type: string
          enum: [pending, processing, completed, failed]
        recordsProcessed:
          type: integer
        recordsFailed:
          type: integer
        errorLog:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time
          nullable: true
...